Separate sqlite / mysql
