+++ .github/workflows/mysql.yml
@@ -86,4 +86,4 @@ jobs:
         run: composer update --prefer-dist --no-interaction --no-progress --optimize-autoloader --ansi
 
       - name: Run tests with phpunit
-        run: vendor/bin/phpunit --colors=always
+        run: vendor/bin/phpunit --group Mysql --colors=always
