+++ tests/Base/TestCase.php
@@ -2,15 +2,8 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Rbac\Cycle\Tests;
+namespace Yiisoft\Rbac\Cycle\Tests\Base;
 
-use Cycle\Database\Config\DatabaseConfig;
-use Cycle\Database\Config\MySQL\DsnConnectionConfig as MyQSLDsnConnectionConfig;
-use Cycle\Database\Config\MySQLDriverConfig;
-use Cycle\Database\Config\Postgres\DsnConnectionConfig as PostgresDsnConnectionConfig;
-use Cycle\Database\Config\PostgresDriverConfig;
-use Cycle\Database\Config\SQLite\FileConnectionConfig;
-use Cycle\Database\Config\SQLiteDriverConfig;
 use Cycle\Database\DatabaseManager;
 use Cycle\Database\Schema\AbstractTable;
 use Symfony\Component\Console\Application;
@@ -72,31 +65,7 @@ abstract class TestCase extends \PHPUnit\Framework\TestCase
         return $app;
     }
 
-    private function createConnection(): void
-    {
-        $dbConfig = new DatabaseConfig(
-            [
-                'default' => 'default',
-                'databases' => [
-                    'default' => ['connection' => 'sqlite'],
-                ],
-                'connections' => [
-                    'sqlite' => new SQLiteDriverConfig(new FileConnectionConfig(__DIR__ . '/runtime/test.db')),
-                    // 'mysql' => new MySQLDriverConfig(new MyQSLDsnConnectionConfig(
-                    //     'mysql:host=127.0.0.1;port=3306;dbname=test',
-                    //     'root',
-                    //     '123456',
-                    // )),
-                    // 'postgres' => new PostgresDriverConfig(new PostgresDsnConnectionConfig(
-                    //     'pgsql:host=127.0.0.1;port=5432;dbname=test',
-                    //     'postgres',
-                    //     '123456',
-                    // )),
-                ],
-            ]
-        );
-        $this->databaseManager = new DatabaseManager($dbConfig);
-    }
+    abstract protected function createConnection(): void;
 
     abstract protected function populateDb(): void;
 }
