+++ tests/Mysql/AssignmentsStorageTest.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Rbac\Cycle\Tests\Mysql;
+
+class AssignmentsStorageTest extends \Yiisoft\Rbac\Cycle\Tests\Base\AssignmentsStorageTest
+{
+    use MysqlTrait;
+}
