+++ tests/Mysql/Command/RbacCycleInitTest.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Rbac\Cycle\Tests\Mysql\Command;
+
+use Yiisoft\Rbac\Cycle\Tests\Mysql\MysqlTrait;
+
+class RbacCycleInitTest extends \Yiisoft\Rbac\Cycle\Tests\Base\Command\RbacCycleInitTest
+{
+    use MysqlTrait;
+}
