+++ tests/Mysql/ItemsStorageTest.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Rbac\Cycle\Tests\Mysql;
+
+class ItemsStorageTest extends \Yiisoft\Rbac\Cycle\Tests\Base\ItemsStorageTest
+{
+    use MysqlTrait;
+}
