+++ tests/Sqlite/Command/RbacCycleInitTest.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Rbac\Cycle\Tests\Sqlite\Command;
+
+use Yiisoft\Rbac\Cycle\Tests\Sqlite\SqliteTrait;
+
+class RbacCycleInitTest extends \Yiisoft\Rbac\Cycle\Tests\Base\Command\RbacCycleInitTest
+{
+    use SqliteTrait;
+}
