+++ tests/Sqlite/SqliteTrait.php
@@ -1,0 +1,31 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Rbac\Cycle\Tests\Sqlite;
+
+use Cycle\Database\Config\DatabaseConfig;
+use Cycle\Database\Config\SQLite\FileConnectionConfig;
+use Cycle\Database\Config\SQLiteDriverConfig;
+use Cycle\Database\DatabaseManager;
+
+trait SqliteTrait
+{
+    protected function createConnection(): void
+    {
+        $dbPath = __DIR__ . DIRECTORY_SEPARATOR . 'runtime' . DIRECTORY_SEPARATOR . 'test.db';
+        $dbConfig = new DatabaseConfig(
+            [
+                'default' => 'default',
+                'databases' => [
+                    'default' => ['connection' => 'sqlite'],
+                ],
+                'connections' => [
+                    'sqlite' => new SQLiteDriverConfig(new FileConnectionConfig($dbPath)),
+                ],
+            ]
+        );
+
+        $this->databaseManager = new DatabaseManager($dbConfig);
+    }
+}
